Added: -context specific decoders for all of the primitive times to go with the context specific encoders -unconfirmed/confirmed EventNotificationRequest structure/encoder/decoder -BACnetTimestamp structure/encoder/decoder -BACnetPropertyStates structure/encoder/decoder -BACnetDeviceObjectPropertyReference structure/encoder/decoder
Changes:
-decode_context_object_id::object_type changed to uint16_t
-explicit casts have been added to some functions to remove compiler warnings
-encode_bacnet_date::year behaviour has been changed slightly

+++ bacnet-stack/src/bacstr.c
@@ -96,10 +96,10 @@ uint8_t bitstring_bits_used(
 }
 
 /* returns the number of bytes that a bit string is using */
-int bitstring_bytes_used(
+uint8_t bitstring_bytes_used(
     BACNET_BIT_STRING * bit_string)
 {
-    int len = 0;        /* return value */
+    uint8_t len = 0;        /* return value */
     uint8_t used_bytes = 0;
     uint8_t last_bit = 0;
